requirements.txt
tkinterdnd2~=0.3.0
cryptography~=42.0.8
PyPDF2~=3.0.1
mutagen~=1.47.0
Pillow~=10.3.0
numpy~=2.0.0